Add install-delay-infrastructure command
